instrumentation/opentelemetry-instrumentation-graphql-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opentelemetry-instrumentation-graphql-core"
dynamic = ["version"]
description = "OpenTelemetry GraphQL-core instrumentation"
readme = "README.rst"
license = "Apache-2.0"
requires-python = ">=3.7"
authors = [
  { name = "OpenTelemetry Authors", email = "cncf-opentelemetry-contributors@lists.cncf.io" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]
dependencies = [
  "opentelemetry-api ~= 1.12",
  "opentelemetry-instrumentation == 0.39b0.dev",
  "opentelemetry-semantic-conventions == 0.39b0.dev",
]

[project.optional-dependencies]
instruments = [
  "graphql-core ~= 3.0",
]
test = [
  "opentelemetry-instrumentation-graphql-core[instruments]",
  "opentelemetry-test-utils == 0.39b0.dev",
]

[project.entry-points.opentelemetry_instrumentor]
graphql-core = "opentelemetry.instrumentation.graphql_core:GraphQLCoreInstrumentor"

[project.urls]
Homepage = "https://github.com/open-telemetry/opentelemetry-python-contrib/tree/main/instrumentation/opentelemetry-instrumentation-graphql-core"

[tool.hatch.version]
path = "src/opentelemetry/instrumentation/graphql_core/version.py"

[tool.hatch.build.targets.sdist]
include = [
  "/src",
  "/tests",
]

[tool.hatch.build.targets.wheel]
packages = ["src/opentelemetry"]
